Add docker-entrypoint
Add docker entry-point and make it executable.
Set the default cmd to server.

The entrypoint fixes permission issues by setting /usr/local/nginx/html/
to belong to user and group www-data.

Always update composer to the most recent one.

Check if docker-entrypoint is invoked with 'server' as argument. If so,
      start and monitor all processes via honcho.
OR try to run the arguments as if they were a cmd line instruction.