Fix permission issues on mounted volumes
Add fix-permission script and run it from docker-entrypoint.

This script checks the UID and GID of a specified path (usually a
mounted volume) and creates a specified user and group with matching
UID and GID.

If the UID/GID of the specified user is already occupied by another
entity. We'll try to remove this entity first.

If the specified user/group already exists, we will try to update the
UID/GID to match the mounted volume. This should fix most of the file
permission issues on mounted volumes.